added rec schema
