+++ src/allotropy/allotrope/models/shared/definitions/custom.py
@@ -10,6 +10,7 @@ from allotropy.allotrope.models.shared.definitions.units import (
     Centimeter,
     Counts,
     CountsPerMilliliter,
+    CountsPerSecond,
     CountsTimesSecond,
     CubicMillimeter,
     DegreeCelsius,
@@ -51,6 +52,8 @@ from allotropy.allotrope.models.shared.definitions.units import (
     RelativeFluorescenceUnitTimesMilliliter,
     RelativeFluorescenceUnitTimesSecond,
     RelativeLightUnit,
+    RelativeLightUnitTimesMilliliter,
+    RelativeLightUnitTimesSecond,
     SecondTime,
     SquareCentimetersPerGram,
     TODO,
@@ -98,6 +101,16 @@ class TQuantityValueCountsPerMilliliter(CountsPerMilliliter, TQuantityValue):
 class TNullableQuantityValueCountsPerMilliliter(
     CountsPerMilliliter, TNullableQuantityValue
 ):
+    pass
+
+
+@dataclass(frozen=True, kw_only=True)
+class TQuantityValueCountsPerSecond(CountsPerSecond, TQuantityValue):
+    pass
+
+
+@dataclass(frozen=True, kw_only=True)
+class TNullableQuantityValueCountsPerSecond(CountsPerSecond, TNullableQuantityValue):
     pass
 
 
@@ -563,6 +576,34 @@ class TQuantityValueRelativeLightUnit(RelativeLightUnit, TQuantityValue):
 @dataclass(frozen=True, kw_only=True)
 class TNullableQuantityValueRelativeLightUnit(
     RelativeLightUnit, TNullableQuantityValue
+):
+    pass
+
+
+@dataclass(frozen=True, kw_only=True)
+class TQuantityValueRelativeLightUnitTimesMilliliter(
+    RelativeLightUnitTimesMilliliter, TQuantityValue
+):
+    pass
+
+
+@dataclass(frozen=True, kw_only=True)
+class TNullableQuantityValueRelativeLightUnitTimesMilliliter(
+    RelativeLightUnitTimesMilliliter, TNullableQuantityValue
+):
+    pass
+
+
+@dataclass(frozen=True, kw_only=True)
+class TQuantityValueRelativeLightUnitTimesSecond(
+    RelativeLightUnitTimesSecond, TQuantityValue
+):
+    pass
+
+
+@dataclass(frozen=True, kw_only=True)
+class TNullableQuantityValueRelativeLightUnitTimesSecond(
+    RelativeLightUnitTimesSecond, TNullableQuantityValue
 ):
     pass
