+++ src/allotropy/allotrope/schemas/shared/definitions/custom.json
@@ -79,6 +79,26 @@
       }
     ]
   },
+  "tQuantityValueCountsPerSecond": {
+    "allOf": [
+      {
+        "$ref": "#/$defs/tQuantityValue"
+      },
+      {
+        "$ref": "#/$defs/CountsPerSecond"
+      }
+    ]
+  },
+  "tNullableQuantityValueCountsPerSecond": {
+    "allOf": [
+      {
+        "$ref": "#/$defs/tNullableQuantityValue"
+      },
+      {
+        "$ref": "#/$defs/CountsPerSecond"
+      }
+    ]
+  },
   "tQuantityValueCountsTimesSecond": {
     "allOf": [
       {
@@ -896,6 +916,46 @@
       },
       {
         "$ref": "#/$defs/RelativeLightUnit"
+      }
+    ]
+  },
+  "tQuantityValueRelativeLightUnitTimesMilliliter": {
+    "allOf": [
+      {
+        "$ref": "#/$defs/tQuantityValue"
+      },
+      {
+        "$ref": "#/$defs/RelativeLightUnitTimesMilliliter"
+      }
+    ]
+  },
+  "tNullableQuantityValueRelativeLightUnitTimesMilliliter": {
+    "allOf": [
+      {
+        "$ref": "#/$defs/tNullableQuantityValue"
+      },
+      {
+        "$ref": "#/$defs/RelativeLightUnitTimesMilliliter"
+      }
+    ]
+  },
+  "tQuantityValueRelativeLightUnitTimesSecond": {
+    "allOf": [
+      {
+        "$ref": "#/$defs/tQuantityValue"
+      },
+      {
+        "$ref": "#/$defs/RelativeLightUnitTimesSecond"
+      }
+    ]
+  },
+  "tNullableQuantityValueRelativeLightUnitTimesSecond": {
+    "allOf": [
+      {
+        "$ref": "#/$defs/tNullableQuantityValue"
+      },
+      {
+        "$ref": "#/$defs/RelativeLightUnitTimesSecond"
       }
     ]
   },
