+++ src/allotropy/allotrope/schemas/shared/definitions/units.json
@@ -47,6 +47,18 @@
       "unit"
     ]
   },
+  "CountsPerSecond": {
+    "properties": {
+      "unit": {
+        "type": "string",
+        "const": "Counts/s",
+        "$asm.unit-iri": "http://purl.allotrope.org/ontology/qudt-ext/unit#CountsPerSecond"
+      }
+    },
+    "required": [
+      "unit"
+    ]
+  },
   "CountsTimesSecond": {
     "properties": {
       "unit": {
@@ -148,7 +160,7 @@
       "unit": {
         "type": "string",
         "const": "μL",
-        "$asm.unit-iri": "http://qudt.org/vocab/unit#Microliter"
+        "$asm.unit-iri": "http://purl.allotrope.org/ontology/qudt-ext/unit#Microliter"
       }
     },
     "required": [
@@ -533,6 +545,30 @@
         "type": "string",
         "const": "RLU",
         "$asm.unit-iri": "http://purl.allotrope.org/ontology/qudt-ext/unit#RelativeLightUnit"
+      }
+    },
+    "required": [
+      "unit"
+    ]
+  },
+  "RelativeLightUnitTimesMilliliter": {
+    "properties": {
+      "unit": {
+        "type": "string",
+        "const": "RLU.mL",
+        "$asm.unit-iri": "http://purl.allotrope.org/ontology/qudt-ext/unit#RelativeLightUnitTimesMilliliter"
+      }
+    },
+    "required": [
+      "unit"
+    ]
+  },
+  "RelativeLightUnitTimesSecond": {
+    "properties": {
+      "unit": {
+        "type": "string",
+        "const": "RLU.s",
+        "$asm.unit-iri": "http://purl.allotrope.org/ontology/qudt-ext/unit#RelativeLightUnitTimesSecond"
       }
     },
     "required": [
